Specified trigger type of Azure Function
When you create a new Azure Function, you have to specify what trigger it should be based on. There is an IoT Hub trigger as well, however the sample code relies on Event Grid trigger. That's why would be very helpful to specify trigger type to avoid later confusion with the wrong trigger execution.