Update waf-front-door-tutorial-geo-filtering.md
As part of this feature (https://msazure.visualstudio.com/One/_workitems/edit/15303787) we're adding SocketAddr as an option for geo filtering. Once this change ships, SocketAddr will be the default for this type of rule since it is a more reliable indicator of where traffic is coming from. We should update this page to instruct users to try building this rule using SocketAddr instead of RemoteAddr. This functionality is already available via API, so we don't need to wait for the portal item to ship before doing this change.