Update apache-spark-autoscale.md
Proposed a clarification based on the example of auto-scaling discussed by the Synapse Discussion team( email group).
The example has the purpose to bring clarification to the configuration that sometimes causes doubts in customers. ( email follow to the page owner)
"In this example, if your job only requires 2 executors, it will only use 2 executors. When the job requires more, it will scale up to 6 executors (1 driver, 5 executors). When the job does not need the executors, then it will decommission the executors and if it does not need the node, it will free up the node.