Global region is not supported for resource groups
If you choose to create a new resource group and choose "global" for resource region, validation will fail because Azure portal try to create the resource group in the same region as the resource but global is not supported for resource groups. So users who are going to create a new resource group should not choose "global".