Grammar
Apologies for the minutiae in advance. :) Never heard this use of "they're". I couldn't find a source here but as a native speaker, I think if you have a clause ending in "they're" you remove the conjunction and just make it "they are". 
It seemed unnatural so I'm proposing this to avoid confusion.