Power Automate Service Dependency on Power Apps.
Power Automate has a service dependency on Power Apps which is documented here (https://docs.microsoft.com/en-us/power-automate/ip-address-configuration#required-services). If a customer is blocking Power Apps, Power Automate will be blocked as well. It's important to add this in the documentation because this did not cause a block to Power Automate until the recent changes to the Power Automate portal (https://powerautomate.microsoft.com/en-us/blog/flow-microsoft-com-is-moving-to-make-powerautomate-com/)