Incorporated changes suggested by Cody and Connor
My team member Barbara Seldon has left the team and the PR was created and assigned to her in May 2022. I am the new owner of this doc, and I was unable to sign-off the PR that was previously created by Cody and Connor. They both are the doc SMEs and original authors. The content is being reviewed and approved by the authors and SMEs.