"Run the app" section will not work as intended
Unless you have all the packages for NuGet installed. I think this wouldn't be a great assumption to make. Or add this to the Prerequisites.

Instead of `Ctrl+F5`, the User should be encouraged to run debugging as this will throw an error if you're missing the package. Also added a note for the users who started from 0 and did the prereqs and downloaded the sample as stated only in this doc.