Azure does not guarantee the behavior
When we use NSG on both subnet and NIC, sometimes we have to check which NSG flow logs are logged. Because Azure does not guarantee the behavior.
In this document, the last sentence is very confusing because we don't know which NSG is processed last. I hope you enhance this document.

https://docs.microsoft.com/en-us/azure/network-watcher/network-watcher-nsg-flow-logging-overview#best-practices
-----
Having NSG at both NIC and Subnet Level: In case NSG is configured at the NIC as well as the subnet level, then flow logging must be enabled at both the NSGs since the exact sequence of rule processing by NSGs at NIC and subnet level is platform dependent and varies from case to case. Traffic flows will be logged against the NSG which is processed last.
-----