Clarify contents of symmetric key if a URI to a file is given.
Putting the base64-encoded derived key generated by the example into a file does not work, and the logged error message is useless. However that's what a user would expect, since they are told to generate a derived key using the Powershell example, and told they can put it in a file.

I clarified that if a file URI is the preferred approach, the key should be stored in binary form.