(AzureCXP) MicrosoftDocs/azure-docs#98678
Changed line 45 to correct the information.
From : VM3: If NSG2 has a security rule that denies port 80, the traffic is denied. If NSG2 has a security rule that allows port 80, then port 80 is allowed outbound to the internet, since a network security group isn't associated to Subnet2.

To : - VM3: If NSG2 has a security rule that denies port 80, the traffic is denied. If not, the traffic is allowed by the AllowInternetOutbound default security rule in NSG2.

Fixes MicrosoftDocs/azure-docs#98678