Update monitor-instances-health-check.md
added the 60s timeout: if the path does not return a response within 1 minute the health check ping is considered unhealthy.

### What happens if my app is running on a single instance?

If your app is only scaled to one instance and becomes unhealthy, it will not be removed from the load balancer because that would take down your application entirely. However, after one hour of returning unhealthy pings, the instance is replaced. Scale out to two or more instances to get the rerouting benefit of Health check. If your app is running on a single instance, you can still use Health check's [monitoring](#monitoring) feature to keep track of your application's health.

### Is HealthCheck following the application code configured redirects between the default domain and the custom domain?

No, Health Check feature is pinging the path of the default domain of the web application. If there is a redirect from the default domain to a custom domain, then the status code that health check is returning is not going to be a 200 but a redirect (301), which is going to mark the worker unhealthy.