Updated enterprise applications deletion steps
Currently, there are few enterprise applications that cannot be deleted in the Azure portal. Engineering is aware of the issue and is actively working on the fix.

In the interim, we have come up with step-by-step instructions on how a customer attempting to delete an Azure AD tenant can unblock themselves today. The link to this document section will be embedded within the error message notification shown in Azure portal. This will help customers navigate to the self-help documentation and follow the steps to be able to delete the tenant.