this applies not only to long running but also to newly created clusters
“… Due to this limitation, if you are planning to use long running clusters (e.g. more than 60 days), we recommend to use the same managed identity for all of the above scenarios …”

From our experience we have noticed in the presented scenarios that the issue occurs not only with long running cluster but also with newly created ones. Therefore, the wording is misleading. So, we propose to remove the highlighted bit. As we think that will bring more clarity to the documentation.