Update multiline-secrets.md
The multiline secret is returned as `My\nSecret`, where the `\n` is NOT the newline character, but rather `\` and `n`. This is unexpected, because the documentation says newline, and I as .NET developer where `\n` is commonly used to denote the invisible newline character expected that those are literal newlines.