Update reset-redemption-status.md
I made some clarification on the necessary privileged roles needed. 
Added a specific section on this and removed it where it was mentioned but not described in full detail. (line 39)

The changes are also technically confirmed by this previous issue/clarification.
#80334

I chose not to include the Directory Writers Role as this role "should not be assigned to any users."
https://learn.microsoft.com/en-us/azure/active-directory/roles/permissions-reference#directory-writers