Update deploy-disaster-recovery-using-jetstream.md
Hello, I am the Lead Solution Architect with JetStream Software. We are in the midst of a documentation improvement effort. I have been tasked with reconciling documentation that is maintained in different location with the current product documentation. The changes I have in this request are the first steps for the AVS docs.