Mentioning why operations may fail if DFS PE is missing
Propose to add below text in this document as while working with a Customer(Case#2210060030001756), they informed that the information already provided in this document is not explicit and they couldn't understand that which operations might fail if they don't add both Blob and DFS endpoint. And also, it's a more generalized information given in this document. So they requested to update this document, if possible, with specific information.

Proposed text(we can mention more operations here):

Similarly, while working with Data Lake Storage Gen2 storage account, if you have added a private endpoint for Blob only and not for DFS it may result in failure of certain operations like Manage ACL, Create Directory, Delete Directory etc as most of the Gen2 operations honor Gen2 APIs and thus it needs a DFS private endpoint.