Update connector-sharepoint-online-list.md
Should App Domain and Redirect URL point to a domain under the implementer's control? Showing contoso.com in the example should help clarify the type of data required. 
Using localhost.com in the example is confusing and potentially irresponsible, as that is a valid domain seemingly not controlled by Microsoft.
https://www.whois.com/whois/localhost.com