Add Warning about failures due to minAPIVersion
A common cause for customer's opening support cases is if they set minAPIVersion property of API Management to a very recent version, newer than what we tell az mon to use to connect to us when creating a diagnostic settings object. This results in a failure and the error message is not very descriptive. Az Mon team has indicated they can't change the error message reported back to users, hence the doc update is being submitted.