Use blob sync instead of upload-batch
In the current workflow changes to the source files are not uploaded using upload-batch, because --overwrite defaults to false. Additionally, if folders or files are removed from the source in GitHub, those will not be removed in the storage account when upload-batch is used. Therefore I propose to use sync instead of upload-batch because sync removes folders and files from the storage account when they are removed from the source in GitHub. Additionally, I suggest to exclude folders and files which should not be internet-facing from uploading (e.g. /.git and /.github folders, and README and LICENSE files).