(AzureCXP) https://github.com/MicrosoftDocs/azure-docs/issues/101875
Adding two FAQs.:

1.	Where are the automatic backups stored?
A.	Automatic backups are simple and stored in the same datacenter as the App Service, and should not be relied upon as your disaster recovery plan.
 
2.	How do I stop the automatic backup?
A.	You cannot stop automatic backup. The automatic backup is stored on the platform and has no effect on the underlying app instance or it’s storage.