Update how-to-connect-group-writeback-enable.md
There's ambiguity in the documentation and what to expect when enabling group writeback. It is not clearly documented in which order changes to default behavior should be done and which order new and original versions need to be enabled or which is the outcome if these orders are not followed.
This ambiguity is generating various support cases created by customers.