Use managed VNET for Azure CNI overlay
The instructions for creating a cluster with Azure CNI overlay say
to create a VNET and subnet. However, Azure CNI overlay supports
managed VNETs, so this step is unnecessary.

Simplify the instructions by removing the VNET/subnet creation
and --vnet-subnet-id parameter for Azure CNI overlay clusters.

Also remove the --node-count parameter in the Cilium instructions
for consistency with the Azure CNI overlay documentation.
The default value of three nodes is sufficient.