Remove previously added sentence about TRIM/UNMAP and managed disks
The Azure Storage PM team has raised an objection to a sentence I added a few days ago.   They don't want to discourage customers from using TRIM/UNMAP because it helps with the overall efficiency and cost of storage.  I'm OK with removing the sentence because the information is already implied by the previous sentence.