Note on Swap With Preview
Added a note to the "Swap with preview" section to mention that the feature is not available if authentication is enabled on one of the slots. There is an error being thrown if a user attempts to start a start with preview (from the UI or Azure DevOps pipeline) that reads: "Conflict - Swap with Preview cannot be used when one of the slots has site authentication enabled. (CODE: 409)". This fact was not documented anywhere and I lost some time re-working my pipeline to start using "swap with preview" and then back to not use it when it turned out it's not possible in our case.