Add note about `appgw.environment`
`appgw.environment` is a required parameter for deploying into Sovereign Clouds, but it is not documented on this page. I had to go into the AGIC source code repo and find a troubleshooting doc that identified it. Adding it to the values list and adding a note so that it gets attention.