Update environment-variables.md
Refactors to add `setx` syntax for SPEECH_REGION as well.
https://learn.microsoft.com/en-us/azure/cognitive-services/speech-service/get-started-speech-to-text?tabs=windows%2Cterminal&pivots=programming-language-csharp

I am following this tutorial and it references the SPEECH_REGION environment variable we should have set. Unless I'm missing it, we are missing the step to actually set the SPEECH_REGION environment variable there and then subsequently telling people to rely on it. That's causing my code to fail with a somewhat obscure error.