Update branch for Spring Apps repository
As the Azure Spring Apps Reference Architecture team transitions the contents of the repository to support a Landing Zone Accelerator centric design, this commit changes the source branch for the repository reference to a protected branch in the repository. This change will enable continued development in the main branch without disrupting the referenced artifacts in the Azure documentation.