Update assign-roles-azure-service-principals.md
If customers have multiple EA Billing Accounts and they want to at the EnrollmentReader or EA Purchaser role at EA billing account scope, they need to grant the SPN the specific role for each EA billing accounts separately. The SPN role assignment at EA Billing Account will not inherit to all the EA billing accounts which customer is not aware of and its not mentioned in this document.

The document does not say anything about currently granting SPN EnrollmentReader roles is all under the hood and its not visible/possible on the UX to check or grant any EnrollmentReader to the SPN.

Regarding the SPN enterprise object ID, the document does not have any reference to help customer find the right SPN object ID. When an Application Registration is created in Azure AD, it will have its own object ID which customer should not be using. Instead when the Application Registrations is created, it will create it's own Enterprise App (SPN). The customer should navigate to the App Registration's Enterprise App and use that Object ID while granting the EA roles to the SPN.

Customers are getting confused on they need to generate a unique GUID for each billingRoleAssignmentName. Hence, it would be good to add a note about the same at the beginning.

Signed-off-by: Darshit Shah <16048145+darushah@users.noreply.github.com>