Update to the "essentials" example payload.
Above this is a table of "essentials".  Guaranteed data that will come from an alert trigger.   
Although Configuration items is listed in that table,  it is excluded from the payload example.   

Extremely valuable data for those using this for logic app creation and want the ConfigurationItems data.  

Additionally, this page conflicts with the Example payload found here:  https://learn.microsoft.com/en-us/azure/azure-monitor/alerts/alerts-common-schema where the configurationItems is clearly in place.