Clarifying host ID override when using slots
The propose change aims to make it more clear/explicit that when slot host id collision is happening the proper way to override the host ID is by setting a different host ID on all setting including and most import the production slot. 

If this is not done, when you do a swap operation the same will fail with a generic error saying "Swap failed. Details: Cannot swap site slots for site 'YOU_SITE_NAME' because the 'YOUR_SLOT_NAME' slot did not respond to http ping.".

This behavior was tested on the following environment:
-> Function on Linux Dedicated Plan
-> Function runtime version 4.14.0.0
-> Language dotnet 6.0 (the code used was the hello world template from the Azure Portal. I created a timer trigger function and an HTTP trigger function )