Update script to test the presence of a directory setting
I've received feedback that some directories may already have a directory setting for 'Group.Unified' present.
In that case, we should only update the existing directory setting property "NewUnifiedGroupWritebackDefault = false".
This update adds a condition to test whether a directory setting already exists. If it does, then update the directory setting property. If it doesn't, it'll create the directory setting.