Updated pre-requisite
Customers are opening request because the connector isn't working but they need to create a diagnostic setting in AAD to stream AAD events to sentinel. Once this is configured, they can use the connector in sentinel. If this isn't enabled in AAD they won't be able to configure the Azure Active Directory connector in Sentinel.