Update lifecycle-management-policy-configure.md
There are a lot of support cases created on this question as customer is not aware of this information and there is no info available on any public documents on the below request as well.

There are blobs in this storage account which has soft delete enabled, which means the retention given in the Life cycle management policy with respect to delete blobs will be executed as per the configuration only. However the blobs will be deleted permanently after the soft delete retention expires.