HTTPS Only should be enabled if Easy Auth is used
If the built-in authentication (aka Easy Auth) is used, but HTTPS Only is disabled, the requests to "/admin/warmup", which originate from within the instance, would be responded to with 302, and the warmup trigger won't be invoked. This is typically not straightforward to figure out. The proposed change clarifies this.

I tested with Windows and Linux Elastic Premium plans, and the observation applies to both.