(AzureCXP) https://github.com/MicrosoftDocs/azure-docs/issues/104117
Adding from discussion internally 

You can move an app to another App Service plan, as long as the source plan and the target plan are in the same resource group, geographical region, and of the same OS type. Any change in type such as Windows to Linux or any type that is different from the originating type is not supported.