Update export-modify-registrations-bulk.md
This sample uses an older SDK of Azure Storage: https://www.nuget.org/packages/windowsazure.storage/ which is deprecated and has been identified with bugs.
I have updated the sample with the latest Azure Storage SDK with the same functionality and removed unnecessay functions. Tested the same and it works!