change Note for webspace
I  added a new note to prevent unintended deletion of certificates in App Service. Some people misunderstand that is is only used by an App Service and remove it when it is no longer needed in the specific App Service. This can might cause service impact. 
If you have a better message, please advise.