(AzureCXP) https://github.com/MicrosoftDocs/azure-docs/issues/104929
as stated in the linked Azure doc, by default, the deployment engine assumes that a ZIP package is ready to run as-is and doesn't run any build automation. To enable the same build automation as in a [Git deployment](https://learn.microsoft.com/azure/app-service/deploy-local-git), set the SCM_DO_BUILD_DURING_DEPLOYMENT app setting) to true. (under the section [Enable build automation for ZIP deploy]).