Precise description to a Python sample
All other (Python) samples are tutorials tailored for a specific scenario such as AAD-only or B2C-only. The sample being called out this time is a template that can support both AAD and B2C (based on configuration), and optionally calls a downstream API. It is one template to meet all the needs.