Update quickstart-configure-dev-box-service.md
On line 66 it is a poor customer experience to keep redirecting readers between pages. We should have all available regions stated here as well.

On line 185, there should be a statement or a warning stating that it is not needed to wait for all running checks to be completed to move on with the deployment. These runs can take forever to complete in big enterprises with a lot of Intune or GPO policies and that will impact the deployment experience. It should also be provided information on what health checks are being supported.

Adding a new sentence on line 193 stating that after the attachment of a network connection to the dev center that a new resource group is created to store the dev box virtual machines NICs. This is very important because it seems not to be possible to change the name of this new resource group yet, and this can effect landing zones' RBACs and Azure Policies, therefore breaking environments.