Splits two modules in single code sample
The current sample has two files in the same code sample. If they are split up, you get an error that the logger isn't defined in the second module. 

This change splits them into separate files and asks that the logger be passed in (or a default logger is used)