Update storage-account-key-note-include.md
I am a MSFT employee, I have added a Note section in the document with the changes that needs to be made. Let me know in case of any issue. Both Product Group and Supportability are informed about the change.
We might also have to update the document later on since Product Group is working on changing the "Authentication type" to "Trusted Access" when using system key however we do not have a definite ETA for the same.