Update cache-network-isolation.md
Currently Redis Enterprise only uses the private endpoint for inbound connections. Outgoing connections like export/import, which are Redis Enterprise reaching out to some other resource, are currently not supported for the private endpoint. Hence, Import/Export is also isn't supported.,
This was learned while working as part of support case and hence raising a request to get this updated in the docs as well.