Indicating path to the execution folder
Code snippet in Step 4 fails, as it refers to main.py script under ./src folder (that was used by the original Jupyter notebook). As this example is being used against the whole repo, it's worth to specify that the long path is  ./sdk/python/jobs/single-step/lightgbm/iris/src, for the people who copy/paste the code to make it work.