[Azure CXP]Fixing wording with Enable a GPU
Description of NVIDIA_VISIBLE_DEVICES seems incorrect.
In step 7 under Enable a GPU in a prefabricated NVIDIA module it describes the NVIDIA_VISIBLE_DEVICES value as "a list of your modules on your device, with 0 being the beginning of the list." But this seems incorrect per Nvidia's repo which says that the value represents a list of GPUs that are exposed to the container for use. E.g. In a multi-GPU environment someone can control whether one or more of those GPUs is available to that container.

Resolves: #https://github.com/MicrosoftDocs/azure-docs/issues/105690