Update subscribe-to-graph-api-events.md
I removed the reference to the need for AAD application IDs to be in an allow list for private preview participation. Updated the  sample Microsoft Graph API subscription request to include the header "x-ms-enable-features: EventGrid" which is now the mechanism through which a user can start sending events to EG during the private preview. Finally, at the top I added our email for any questions they may have.