Update create-self-hosted-integration-runtime.md
Based on my most recent experience (2023-02-20) it looks Like JAVA 11 is the currently supported JRE / JDK Now (it is unclear if 8 is still supported or not. I had numerous JNI erros with 8 that seemed to go away after moving to 11). Also depending on the distro the bin folder may not get automatically added to the system path with the JRE install. I have added this as another thing to check. if the bin folder is not in the system path the SHIR will be unable to find the java.exe   Please verify this for technical accuracy thanks