Update_service_account_cmdlet.md
In the section where we are manually enabling the AES 256 encryption, it is mentioned that the storage account can be a computer account or a service logon account.
The powershell cmdlet to enable encryption is mentioned only for a computer account and not a service logon account. This creates some confusion when someone is following the document because if they have a service logon account this command does not work and they think that this command should only be used to change encryption whether it is computer account or service logon account.
So, I think it will be easy for readers to understand if we clearly mention both the cmdlets for different types of account.