Update virtual-machine-scale-sets-automatic-instance-repairs.md
I am working as an Azure Cloud Engineer in Microsoft. Recently I received a customer support request where the customer was unclear about what he needs to mention in the "Path" field while enabling the Health and Repair option for his existing VMSS. When I checked this existing Microsoft document, I found that the exact steps are not mentioned in it and hence tried to update it with more granular/descriptive way.
Please verify from your end and provide your feedback.