Remove support for Private Custom Domain
Private Custom Domains are not supported on AppServices. They used to be with ASEv2, when customers were able to choose their own domain, even if they didn't publicly own the domain, but this ability has since been removed. If a customer wishes to add a domain to their AppService, they must own the domain and it must be publicly resolvable.