(AzureCXP) fixes MicrosoftDocs/azure-docs#106142
Changes made in line no  47
From
On the Settings blade for the storage account, select Encryption. By default, key management is set to Microsoft-managed keys, as shown in the following image.
To
On the Under Security + networking, select Encryption. By default, key management is set to Microsoft-managed keys, as shown in the following image.