Update private-link-faq.yml
Network policies like NSGs (Network security groups) and User-Defined Routes previously weren't supported for private endpoints. To deploy a private endpoint on a given subnet, an explicit disable setting was required on that subnet. This setting is only applicable for the private endpoint. For other resources in the subnet, access is controlled based on security rules in the network security group.
Private Endpoint support for Network Security Groups (NSGs) is now in GA. 
This feature enhancement will provide you with the ability to enable advanced security controls on traffic destined to a private endpoint.
If customer is interested in using this feature, they will need to set a specific subnet level property, called PrivateEndpointNetworkPolicies, to Enabled.