Update bitnami-appliances-deployment.md
We are trying to get rid of instances where we talk about SSH using PuTTY. PowerShell supports OpenSSH, so it is available for Windows and mac\Linux both have Bash. I think this syntax is correct, but you might want to try it an make sure the user isn't prompted for anything else. Thank you!