Change to the term Cores from Nodes
In the Documentation, there is no clarity of what is a core But it is using the term nodes everywhere to represent the number of instances that it will run. So I believe, even the table titles must be nodes, not cores, as I couldn't find any definition of cores in the IR documentation at all.