Line no 72 can be improved.
One of my customers reach out to me about the below statement on SAS expiry time. As per my research, there's no maximum time limit imposed by Azure on the expiry of a SAS URL. We can set it to 9999-12-31T23:59:59Z so that it never expires. 

"The value for the expiry time is a maximum of seven days from the creation of the SAS token"