Wrong names for our own products
This page referred to "Defender for Azure Resource Manager". 
The plan does not have the word "Azure" in its name. 
If unsure, check [the product docs](https://learn.microsoft.com/en-us/azure/defender-for-cloud/defender-for-resource-manager-introduction). 

Worse, this page referred to "Defenders for Cloud". It is "Defender", singular.

Lastly, the page has inconsistencies with regards to the name of our servers plan:
Once, it used the singular "Defender for Server" and other places (the correct) "Defender for Servers"

-------
cc: @ElazarK