Create gateway-IP-addresses.md
This page is needed, so customers are not referenced to https://learn.microsoft.com/en-us/azure/azure-sql/database/connectivity-architecture?view=azuresql. Not all parts of that page apply to Synapse Analytics workspace, or Synapse Analytics Serverless, so we cannot add those to "Applies to" section. On the other hand, we cannot move Gateway IP addresses to its own page, since some resources are already referencing the existing page.