Remove deprecated configuration options
The configuration steps related to the input fields labeled "will be deprecated" are no longer required. These configuration options (Service Provider ID, IdP SSO Service URL, Do Not Deflate SP-initiated Authentication Request) are now present in the SAML Identity Providers grid described in the previous step. These options are specific to each individual SAML identity provider and do no need to be specified here. 

The Authentication Request Signature Method should default to SHA256 and does not need to be specified here. We can create a separate PR for the x509 Private Key Pair, which is used for logouts. This is the main functionality I've previously configured in this section, but more information would be required on how to create the private key pair within Workday and loading the public portion to Azure.