changed <pre><code> to    ```bash around code
Changed <pre><code> to    ```bash around code and removed bolding.  The spaces in front of the actual code caused a problem with the "here document", because the last line had a space before the EOF. This caused this to not work properly.  Also, in the second line of the document, there was a typeo, where it said hni rather than hn1.  this was corrected in this PR.