Urgent: CX contacts support this issue often.
For the below MS Doc, that would be helpful if there is an explanation that SaaS subscription is not stopping charge if users cancel Azure subscription. SaaS subscription is different that Azure subscription. For example, users purchase SendGrid service through Azure Marketplace, and it means that users are not purchasing Azure service. This section explains only Azure service, and it is not Marketplace product which is SaaS subscription. Because of the documentation, users misunderstood if they cancel Azure subscription means that SaaS subscription also stops charging. It is not correct understanding, and it would be charging even though they think that SaaS subscription has been canceled.

What happens after subscription cancellation?
https://learn.microsoft.com/en-us/azure/cost-management-billing/manage/cancel-azure-subscription?source=recommendations#what-happens-after-subscription-cancellation

I have created a request below. I have not heard any respond in the request.  
https://github.com/MicrosoftDocs/azure-docs/issues/106507