[Doc-a-thon] Updating cloudinit-configure-swapfile
Removing the reference to sensible-editor, given that's not installed in all Linux distros and the sentence is already mentioning they can use the text editor of their preference. Removing also the centos7 and URN image from the sample entry for the vm create to make it Linux agnostic (they could use any image with Cloud-init), changing the last block code from bash to output. I also added sudo to the commands