(AzureCXP) fixes MicrosoftDocs/azure-docs#107447
Updated line number 15
From:
"The amount of egress data. This number includes egress from an external client into Azure Storage as well as egress within Azure. As a result, this number does not reflect billable egress."

To:
"The amount of egress data. This number includes egress to an external client from Azure Storage as well as egress within Azure (between Azure storage/services). As a result, this number does not reflect billable egress."